+++ packages/backend/src/api/resources/JSON.ts
@@ -246,6 +246,8 @@ export interface SamlAccountJSON extends ClerkResourceJSON {
   last_name: string;
   verification: VerificationJSON | null;
   saml_connection: SamlAccountConnectionJSON | null;
+  last_authenticated_at: number | null;
+  enterprise_connection_id: string | null;
 }
 
 export interface IdentificationLinkJSON extends ClerkResourceJSON {