+++ packages/clerk-js/src/core/resources/BillingCheckout.ts
@@ -13,6 +13,7 @@ import type {
 import { unixEpochToDate } from '@/utils/date';
 
 import { billingTotalsFromJSON } from '../../utils';
+import { Billing } from '../modules/billing/namespace';
 import { BillingPayer } from './BillingPayer';
 import { BaseResource, BillingPaymentMethod, BillingPlan } from './internal';
 
@@ -62,9 +63,7 @@ export class BillingCheckout extends BaseResource implements BillingCheckoutReso
     return retry(
       () =>
         this._basePatch({
-          path: this.payer.organizationId
-            ? `/organizations/${this.payer.organizationId}/commerce/checkouts/${this.id}/confirm`
-            : `/me/commerce/checkouts/${this.id}/confirm`,
+          path: Billing.path(`/checkouts/${this.id}/confirm`, { orgId: this.payer.organizationId }),
           body: params as any,
         }),
       {