+++ packages/clerk-js/src/ui/customizables/elementDescriptors.ts
@@ -102,6 +102,7 @@ export const APPEARANCE_KEYS = containsAllElementsConfigKeys([
   'otpCodeFieldInput',
   'otpCodeFieldInputContainer',
   'otpCodeFieldErrorText',
+  'otpCodeFieldSuccessText',
   'formResendCodeLink',
 
   'dividerRow',
@@ -505,6 +506,10 @@ export const APPEARANCE_KEYS = containsAllElementsConfigKeys([
   'subscriptionDetailsDetailRow',
   'subscriptionDetailsDetailRowLabel',
   'subscriptionDetailsDetailRowValue',
+
+  'enterpriseConnectionsRoot',
+  'enterpriseConnectionButton',
+  'enterpriseConnectionButtonText',
 ] as const).map(camelize) as (keyof ElementsConfig)[];
 
 type TargettableClassname<K extends keyof ElementsConfig> = `${typeof CLASS_PREFIX}${K}`;