+++ packages/shared/src/errors/parseError.ts
@@ -1,34 +1,25 @@
-import type { ClerkAPIError, ClerkAPIErrorJSON } from '@clerk/types';
+import type { ClerkAPIError as ClerkAPIErrorInterface, ClerkAPIErrorJSON } from '@clerk/types';
+
+import { ClerkAPIError } from './clerkApiError';
 
 /**
  * Parses an array of ClerkAPIErrorJSON objects into an array of ClerkAPIError objects.
  *
  * @internal
  */
-export function parseErrors(data: ClerkAPIErrorJSON[] = []): ClerkAPIError[] {
-  return data.length > 0 ? data.map(parseError) : [];
+export function parseErrors(data: ClerkAPIErrorJSON[] = []): ClerkAPIErrorInterface[] {
+  return data.length > 0 ? data.map(e => new ClerkAPIError(e)) : [];
 }
 
 /**
  * Parses a ClerkAPIErrorJSON object into a ClerkAPIError object.
  *
+ * @deprecated Use `ClerkAPIError` class instead
+ *
  * @internal
  */
-export function parseError(error: ClerkAPIErrorJSON): ClerkAPIError {
-  return {
-    code: error.code,
-    message: error.message,
-    longMessage: error.long_message,
-    meta: {
-      paramName: error?.meta?.param_name,
-      sessionId: error?.meta?.session_id,
-      emailAddresses: error?.meta?.email_addresses,
-      identifiers: error?.meta?.identifiers,
-      zxcvbn: error?.meta?.zxcvbn,
-      plan: error?.meta?.plan,
-      isPlanUpgradePossible: error?.meta?.is_plan_upgrade_possible,
-    },
-  };
+export function parseError(error: ClerkAPIErrorJSON): ClerkAPIErrorInterface {
+  return new ClerkAPIError(error);
 }
 
 /**